feat: add page header and style
This commit adds a page header to the application and styles it using UnoCSS.

The page header includes a logo, navigation links, and a button. The navigation links are styled to be underlined when hovered over, and the button is styled with a gradient background.

UnoCSS is a utility-first CSS framework that makes it easy to add styles to your application. It provides a set of pre-defined classes that you can use to style your elements. In this commit, we use UnoCSS to style the page header and the button.